requirements.txt
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
gspread==5.12.0
gspread-dataframe==3.3.1
requests
pygsheets
pandas
numpy
holidays
pytorch-forecasting==1.0.0
tensorboard==2.15.1
tensorboardX==2.6.2.2
torch==2.1.1
torchaudio==2.1.1
torchmetrics==1.2.1
torchvision==0.16.1
cpflows==0.1.2